pyproject.toml
[project]
name = "vad-benchmark"
version = "0.1.0"
description = "Local reproduction of Silero-VAD quality metrics, extended with AI-coustics Quail VAD"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "soundfile>=0.12",
    "resampy>=0.4",
    "scikit-learn>=1.4",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "rich>=13.7",
]

[project.optional-dependencies]
webrtc = ["webrtcvad-wheels>=2.0.14"]
silero = ["silero-vad>=6.0", "torch>=2.2"]
aicoustics = ["aic-sdk>=2.0"]
all = [
    "webrtcvad-wheels>=2.0.14",
    "silero-vad>=6.0",
    "torch>=2.2",
    "aic-sdk>=2.0",
]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
vad-bench = "vad_benchmark.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vad_benchmark"]

[tool.ruff]
line-length = 100
target-version = "py310"
